// gen/ts/src/common/types.ts
export enum ConnectorTransport {
  CONNECTOR_TRANSPORT_UNSPECIFIED = "CONNECTOR_TRANSPORT_UNSPECIFIED",
  UNIX_SOCKET = "UNIX_SOCKET",
  MTLS_TCP = "MTLS_TCP",
  UNRECOGNIZED = "UNRECOGNIZED",
}